Give ESMFold one config file and an environment that stands alone
pyproject.toml is the whole manifest now: requirements.txt held a single line it
read back through dynamic dependencies, and the deps it did not name (torch,
numpy) were left to arrive as somebody else's transitive.

The environment brings its own Python. It was a venv on whatever `python3` the
host had, which on a cluster login node is routinely older than the package's
own requires-python, so the install could only fail there. micromamba already
provisions the OpenFold environment and the dashboard's Node one, so `mamba::ensure`
in the shared lib now serves all three.

The entrypoint moved into the package as `esmfold.cli`, installed as
`vizfold-esmfold`: folding no longer reaches back into the checkout for a script,
and `scripts/esmfold/run_pretrained_esmf.py` calls the same function for anyone
running it by path. `status` checks the environment has it, and the install
verifies the whole chain -- interpreter, every import, entrypoint -- rather than
that pip exited zero.
