+++ backends/esmfold/install/install.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
-# Install the ESMFold backend: a plain venv with PyTorch + Transformers. No SLURM/site machinery
-# and no AF2 databases -- ESMFold needs no CUDA build and pulls its weights from HuggingFace at
-# run time. Invoked by `vizfold install esmfold`. Idempotent: skips the pip work if the env
-# already imports torch + transformers.
+# Install the ESMFold backend into an environment that can fold on its own: its own Python, PyTorch,
+# Transformers, and the `esmfold` package with its `vizfold-esmfold` entrypoint. No SLURM/site
+# machinery and no AF2 databases -- ESMFold needs no CUDA build and pulls its weights from
+# HuggingFace at run time. Invoked by `vizfold install esmfold`. Idempotent: skips the pip work if
+# the environment already imports what it needs.
 set -euo pipefail
 
 # config.sh is the backend-neutral shared install lib (repo-root lib/), owned by no backend. It
@@ -18,33 +19,47 @@ REPO=${OPENFOLD_HOME:-$REPO}
 ESM=$REPO/backends/esmfold
 PREFIX=${OPENFOLD_PREFIX:-$HOME/openfold}
 ENV=${ESMFOLD_ENV_PREFIX:-$(vizfold::env esmfold)}
+# The package needs >=3.10, which a cluster login node's python3 routinely is not -- so the
+# environment brings its own rather than inheriting whichever one happens to be on PATH.
+PYTHON_VERSION=3.11
 test -f "$ESM/pyproject.toml" || die "no esmfold project at $ESM; is $REPO a vizfold checkout?"
-command -v python3 >/dev/null || die "python3 is required to create the ESMFold venv"
 
 esmfold::present() { "$ENV/bin/python" -c 'import torch, transformers, esmfold' 2>/dev/null; }
 
-esmfold::install() {
-    log "venv $ENV"
+esmfold::env() {
+    log "env $ENV (python $PYTHON_VERSION)"
+    local mm; mm=$(mamba::ensure "$PREFIX")
+    export MAMBA_ROOT_PREFIX=$PREFIX/mamba
     mkdir -p "$(dirname "$ENV")"
-    [ -x "$ENV/bin/python" ] || python3 -m venv "$ENV"
-    "$ENV/bin/python" -m pip install --upgrade pip
-    # torch first (own wheel index, e.g. a CUDA build); transformers from PyPI reads requirements.
+    # --no-rc so a user ~/.condarc envs_dirs/channels cannot redirect it, as the OpenFold install does.
+    "$mm" create -y --no-rc -p "$ENV" -c conda-forge "python=$PYTHON_VERSION" pip
+}
+
+esmfold::install() {
+    # torch first, from its own wheel index when one is set (a CUDA build): pyproject asks only for
+    # torch>=2.1, so the package install below finds it satisfied and leaves this build alone.
     log torch
     local index=()
     [ -n "${ESMFOLD_PIP_INDEX_URL:-}" ] && index=(--index-url "$ESMFOLD_PIP_INDEX_URL")
     "$ENV/bin/pip" install ${index[@]+"${index[@]}"} "${ESMFOLD_TORCH_SPEC:-torch}"
-    # the esmfold project pulls its deps (transformers) from pyproject and installs the `esmfold`
-    # package itself, so `import esmfold` resolves in the venv the run entrypoint executes under.
+    # The project declares the rest and installs the `esmfold` package with its entrypoint.
     log package
     "$ENV/bin/pip" install "$ESM"
 }
 
+# Prove the environment can fold on its own: its own interpreter, every import the backend makes,
+# and the entrypoint on its PATH -- not merely that the pip commands exited zero.
 esmfold::verify() {
     log verify
     "$ENV/bin/python" - <<'PY'
-import torch, transformers
+import shutil, sys, torch, transformers
+from esmfold.cli import main
+print("python", sys.version.split()[0])
 print("torch", torch.__version__, "cuda", torch.cuda.is_available())
 print("transformers", transformers.__version__)
+entrypoint = shutil.which("vizfold-esmfold", path=f"{sys.prefix}/bin")
+assert entrypoint, "vizfold-esmfold is not in the environment"
+print("entrypoint", entrypoint)
 PY
 }
 
@@ -61,6 +76,7 @@ main() {
     if esmfold::present; then
         log "already installed at $ENV"
     else
+        [ -x "$ENV/bin/pip" ] || esmfold::env
         esmfold::install
     fi
     esmfold::verify
@@ -72,7 +88,7 @@ ESMFold env: $ENV
 
 Fold the bundled example (downloads facebook/esmfold_v1 on first run):
 
-  $ENV/bin/python $REPO/scripts/esmfold/run_pretrained_esmf.py \\
+  $ENV/bin/vizfold-esmfold \\
     --fasta $REPO/examples/monomer/fasta_dir_6KWC/6KWC.fasta \\
     --out $PREFIX/outputs/esmf_6KWC --trace_mode none
 EOF
