+++ backends/esmfold/pyproject.toml
@@ -8,14 +8,20 @@ version = "0.1.0"
 description = "ESMFold backend for vizfold: HuggingFace EsmForProteinFolding with attention/activation tracing."
 requires-python = ">=3.10"
 license = { text = "Apache License, Version 2.0" }
-# torch is deliberately absent: install.sh installs it first from its own wheel index (a CUDA
-# build) so the package install below links against it rather than pulling a PyPI copy.
-dynamic = ["dependencies"]
+# Everything the backend imports, so the installed package alone can fold. torch is declared but
+# install.sh installs it first from its own wheel index when a CUDA build is wanted: pip then finds
+# the requirement already satisfied and leaves that build in place.
+dependencies = [
+    "torch>=2.1",
+    "transformers>=4.36.0,<5.0.0",
+    "numpy>=1.24",
+]
 
-[tool.setuptools.dynamic]
-dependencies = { file = ["requirements.txt"] }
+# The environment's own entrypoint, so a fold needs nothing outside it.
+# scripts/esmfold/run_pretrained_esmf.py runs the same function, for calling it by path.
+[project.scripts]
+vizfold-esmfold = "esmfold.cli:main"
 
 [tool.setuptools.packages.find]
-# only the esmfold package; the run_pretrained_esmf.py entrypoint stays a plain script.
 namespaces = false
 include = ["esmfold*"]
